requirements.txt
fastapi 
uvicorn 
pyjwt 
passlib
sqlalchemy 
sqlalchemy-utils 
pymysql 
databases 
alembic
pydantic-settings
aiomysql
python-jose
git+https://github.com/dpkp/kafka-python.git
python-multipart
cryptography
aiokafka
pytz